Split CosmWasmClient and SigningCosmWasmClient
